Refactor code structure and remove unused test files
