+++ tests/test_training.py
@@ -1,5 +1,126 @@
-from agogos.training import ParallelTrainingSystem, Trainer, TrainingSystem
 import pytest
+from agogos.training import Trainer, TrainingSystem, ParallelTrainingSystem
+
+
+class TestTrainer:
+    def test_trainer_abstract_train(self):
+        trainer = Trainer()
+        with pytest.raises(NotImplementedError):
+            trainer.train([1, 2, 3], [1, 2, 3])
+
+    def test_trainer_abstract_predict(self):
+        trainer = Trainer()
+        with pytest.raises(NotImplementedError):
+            trainer.predict([1, 2, 3])
+
+    def test_trainer_train(self):
+        class trainerInstance(Trainer):
+            def train(self, x, y):
+                return x, y
+
+        trainer = trainerInstance()
+        assert trainer.train([1, 2, 3], [1, 2, 3]) == ([1, 2, 3], [1, 2, 3])
+
+    def test_trainer_predict(self):
+        class trainerInstance(Trainer):
+            def predict(self, x):
+                return x
+
+        trainer = trainerInstance()
+        assert trainer.predict([1, 2, 3]) == [1, 2, 3]
+
+    def test_trainer_hash(self):
+        trainer = Trainer()
+        assert trainer.get_hash() == "0a1fcf1d677d4a1f3f082aa85ffcb684"
+
+
+class TestTrainingSystem:
+    def test_training_system_init(self):
+        training_system = TrainingSystem()
+        assert training_system is not None
+
+    def test_training_system_init_with_steps(self):
+        class SubTrainer(Trainer):
+            def predict(self, x):
+                return x
+
+        block1 = SubTrainer()
+        training_system = TrainingSystem(steps=[block1])
+        assert training_system is not None
+
+    def test_training_system_wrong_step(self):
+        class SubTrainer:
+            def predict(self, x):
+                return x
+
+        with pytest.raises(AssertionError):
+            TrainingSystem(steps=[SubTrainer()])
+
+    def test_training_system_steps_changed_predict(self):
+        class SubTrainer:
+            def predict(self, x):
+                return x
+
+        block1 = SubTrainer()
+        training_system = TrainingSystem()
+        training_system.steps = [block1]
+        with pytest.raises(TypeError):
+            training_system.predict([1, 2, 3])
+
+    def test_training_system_predict(self):
+        class SubTrainer(Trainer):
+            def predict(self, x):
+                return x
+
+        block1 = SubTrainer()
+        training_system = TrainingSystem(steps=[block1])
+        assert training_system.predict([1, 2, 3]) == [1, 2, 3]
+
+    def test_trainsys_predict_with_trainer_and_trainsys(self):
+        class SubTrainer(Trainer):
+            def predict(self, x):
+                return x
+
+        block1 = SubTrainer()
+        block2 = SubTrainer()
+        block3 = TrainingSystem(steps=[block1, block2])
+        training_system = TrainingSystem(steps=[block1, block2, block3])
+        assert training_system.predict([1, 2, 3]) == [1, 2, 3]
+
+    def test_training_system_train(self):
+        class SubTrainer(Trainer):
+            def train(self, x, y):
+                return x, y
+
+        block1 = SubTrainer()
+        training_system = TrainingSystem(steps=[block1])
+        assert training_system.train([1, 2, 3], [1, 2, 3]) == ([1, 2, 3], [1, 2, 3])
+
+    def test_traiinsys_train_with_trainer_and_trainsys(self):
+        class SubTrainer(Trainer):
+            def train(self, x, y):
+                return x, y
+
+        block1 = SubTrainer()
+        block2 = SubTrainer()
+        block3 = TrainingSystem(steps=[block1, block2])
+        training_system = TrainingSystem(steps=[block1, block2, block3])
+        assert training_system.train([1, 2, 3], [1, 2, 3]) == ([1, 2, 3], [1, 2, 3])
+
+    def test_training_system_steps_changed_train(self):
+        class SubTrainer:
+            def train(self, x, y):
+                return x, y
+
+        block1 = SubTrainer()
+        training_system = TrainingSystem()
+        training_system.steps = [block1]
+        with pytest.raises(TypeError):
+            training_system.train([1, 2, 3], [1, 2, 3])
+
+    def test_training_system_empty_hash(self):
+        training_system = TrainingSystem()
+        assert training_system.get_hash() == ""
 
 
 class TestParallelTrainingSystem:
